requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.31.0
rich==14.1.0
termcolor==2.3.0
urllib3==2.0.3
